ch07/requirements.txt
# Chapter 7: AI and Advanced Data Transformations - Requirements
# Core data manipulation and analysis
pandas>=2.0.0

# AI and API integration
openai>=1.3.0

# Environment and configuration management
python-dotenv>=1.0.0

# Data validation and modeling
pydantic>=2.0.0

# String matching and fuzzy logic
rapidfuzz>=3.0.0

# Date and time manipulation
pytz>=2023.3
python-dateutil>=2.8.0

# Progress bars and user interface
tqdm>=4.65.0

# Data generation for testing (used in lab exercises)
faker>=19.0.0

# Jupyter notebook support (optional but recommended)
jupyter>=1.0.0
ipykernel>=6.25.0

# Additional utilities that may be useful
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0 